requirements.txt
MarkupSafe==3.0.2
PyJWT==2.10.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
PyYAML==6.0.2
SQLAlchemy==2.0.36
Werkzeug==3.1.3
XlsxWriter==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==4.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b3
azure-ai-formrecognizer==3.3.1
azure-ai-translation-text==1.0.1
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.14.1
azure-search==1.0.0b2
azure-search-documents==11.4.0b11
azure-storage-blob==12.20.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
docstring-parser==0.16
docx2txt==0.8
docx2txt2==1.0.4
entrypoints==0.4
et-xmlfile==2.0.0
faiss-cpu==1.9.0.post1
flask==3.1.0
flask-cors==5.0.0
fonttools==4.55.0
frozenlist==1.5.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.43
langchain-google-vertexai==1.0.10
langchain-text-splitters==0.2.4
langchain_openai==0.1.20
langdetect==1.0.9
langsmith==0.1.146
lxml==5.3.0
marshmallow==3.23.1
matplotlib==3.9.2
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.7
openpyxl==3.1.5
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==10.1.0
portalocker==2.10.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.6.1
pydantic-core==2.16.2
pyparsing==3.2.0
pypdf==4.2.0
pypdf2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
rank_bm25==0.2.2
referencing==0.35.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.21.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toolz==1.0.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vertexai==1.71.1
waitress==3.0.2
xlrd==2.0.1
yarl==1.18.0